[Fix] Fix to component structure - Move info-folder to components folder
