+++ src/App.jsx
@@ -4,7 +4,7 @@ import { MAIN_TITLE } from './utils/constants/constants';
 
 import TodoList from './components/todo/TodoList';
 import TodoInput from './components/input/TodoInput';
-import TodoFilter from './info-bar/TodoFilter';
+import TodoFilter from './components/info-bar/TodoFilter';
 
 const App = () => (
   <>
